timeouts

+++ src/constant/communication/general.py
@@ -6,5 +6,5 @@ TIMEOUT: int = int(_config["timeout"])
 BUFFER_SIZE: int = int(_config["buffer_size"])
 
 # Reliable communication
-RELIABLE_TIMEOUT: int = int(_config["reliable_timeout"])
+RELIABLE_TIMEOUT: float = float(_config["reliable_timeout"])
 RELIABLE_RETRIES: int = int(_config["reliable_retries"])
